refactor: export tracing from engine-core
