+++ packages/engine-core/package.json
@@ -31,7 +31,6 @@
   "dependencies": {
     "@opentelemetry/sdk-trace-base": "^1.4.0",
     "@opentelemetry/api": "^1.1.0",
-    "@opentelemetry/core": "^1.4.0",
     "@prisma/debug": "workspace:*",
     "@prisma/engines": "workspace:*",
     "@prisma/generator-helper": "workspace:*",
