+++ packages/engine-core/src/index.ts
@@ -17,3 +17,4 @@ export { plusX } from './common/utils/util'
 export { DataProxyEngine } from './data-proxy/DataProxyEngine'
 export { LibraryEngine } from './library/LibraryEngine'
 export * as NodeAPILibraryTypes from './library/types/Library'
+export * from './tracing'
