+++ packages/engine-core/src/tracing/index.ts
@@ -1,0 +1,4 @@
+export { createSpan } from './createSpan'
+export { getTraceParent } from './getTraceParent'
+export { runInActiveSpan, runInSpan } from './runInSpan'
+export { TransactionTracer } from './TransactionTracer'
